requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.9
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
pydantic==2.5.0
